requirements-dev.txt
# Include base requirements
-r requirements.txt

# Development and testing dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Additional testing utilities
requests>=2.28.0  # For integration tests
factory-boy>=3.2.0  # For test data generation